fix: Clean up copyright, todo, and why rules.
Update copyright patterns for silver+ tiers to assert that a copyright header exits at all, rather than enforce Outreach copyright.

Update TODO patterns to not for a Jira ticket, be a little more lenient in ticket formatting, and to have some unit tests.

Update why patterns to be a little cleaner and to have unit tests.

Add some very basic documentation to the README.